Min-value - long int

+++ op_functions.c
@@ -78,8 +78,8 @@ int print_int(va_list args)
 int print_dec(va_list args)
 {
 	int count;
-	int i;
-	int n;
+	long int i;
+	long int n;
 
 	count = 0;
 
@@ -112,7 +112,7 @@ int print_dec(va_list args)
 
 }
 
-void print_numbers(int i)
+void print_numbers(long int i)
 {
 	if (i / 10)
 	{
